TESTING WITH DUMMY DATA (POSTS)
Inside blog apps views.py file, we create two dummy posts using dictionary and
store them in a context dictionary and pass it to render in the template. We
then edit home.html by adding some code-block to loop over the posts to show
in the frontend. We haven't used any databases yet.